refactor: reorganiser le code en déplaçant la gestion des builtins et des entrées utilisateur dans de nouveaux fichiers
