--- read_line.c
@@ -1,37 +1,0 @@
-#include "shell.h"
-
-/**
- * print_prompt - Displays the shell prompt
- */
-void print_prompt(void)
-{
-	if (isatty(STDIN_FILENO))
-	{
-		printf("($) ");
-		fflush(stdout);
-	}
-}
-
-/**
- * read_line - Reads a line from stdin
- * Return: pointer to the read string, or NULL if EOF
- */
-char *read_line(void)
-{
-	char *line = NULL;
-	size_t size = 0;
-	ssize_t nread;
-
-	nread = getline(&line, &size, stdin);
-	if (nread == -1)
-	{
-		if (line)
-			free(line);
-		return (NULL);
-	}
-
-	if (nread > 0 && line[nread - 1] == '\n')
-		line[nread - 1] = '\0';
-
-	return (line);
-}
